[3/4/26]: [CudaArrTex] [CudaVirtualMemory] [CudaDeviceMemory] [CudaManagedMemory] [CudaUtils] [CudaUtility]
[Init]:
- Added the CudaArrTex files
- Added the CudaUtility files

[Add]:
  - Added CudaArray, MipmappedArray, TexObject, SurfObject handle structs
  - Added ArrayFormat, ArrayFlags, CopyMemoryType, ResourceType, TexAddressMode, TexFilterMode, TexFlags, ResourceViewFormat, CONCURRENT_MANAGED_ACCESS enums
  - Added Array3dDesc, MemCpy3DDesc, ResourceDesc, TextureDesc, ResourceViewDesc descriptors with full mutator sets
  - Added validateArray3dDesc, validateMemCpy3DDesc, validateResourceDesc, validateTextureDesc, validateResourceViewDesc
  - Added toArrayFormat, toArrayFlags, toResourceType, toTexAddressMode, toTexFilterMode, toResourceViewFormat, toCopyMemoryType
  - Added CUDA_PackingFunctions class with packArray3dDesc, pack3dMemcpyDesc, packResourceDesc, packTextureDesc, packResourceViewDesc

[Fix[:
  - Fixed OOB s_Devices[-1] when a CPU DeviceHandle was passed to adviseMemory
  - Fixed CUDA_ERROR_INVALID_DEVICE in adviseMemory and prefetchAsync by guarding with CONCURRENT_MANAGED_ACCESS check on Windows WDDM
  - Fixed raw m_HandleValue usage in prefetchAsync — now goes through s_Devices registry consistently

+++ src/SpectraCudaBackend/src/Private/CudaDeviceMemory.cpp
@@ -20,7 +20,7 @@ namespace Spectra::Cuda::Memory {
 			return addr;
 	}
 
-	PitchedAllocation DeviceMemory::deviceAllocPitch(size_t v_WidthInBytes, size_t v_Height, size_t v_ElemsInBytes) {
+	PitchedAllocation DeviceMemory::deviceAllocPitch(size_t v_WidthInBytes, size_t v_Height, uint32_t v_ElemsInBytes) {
 		PitchedAllocation alloc{};
 		CUdeviceptr ptr{};
 		size_t pitch = 0;
